Tdep/stripped services

+++ src/main/java/com/tdep/tadlab/repository/LinkRepository.java
@@ -1,13 +1,12 @@
 package com.tdep.tadlab.repository;
 
-import com.tdep.tadlab.entity.Link;
+import com.tdep.tadlab.entity.projectDb.Link;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.UUID;
 
 @Repository
-public interface LinkRepository extends JpaRepository<Link, Long> {
+public interface LinkRepository extends JpaRepository<Link, Integer> {
 
 }
